examples/app-clock-volume-usbclcd1602/requirements.txt
hidapi
pstray
Pillow
pycaw
psutil